@@ -2,7 +2,7 @@
   	<modelVersion>4.0.0</modelVersion>
   	<groupId>io.github.mariazevedo88</groupId>
   	<artifactId>goldgem</artifactId>
-  	<version>1.0.1</version>
+  	<version>2.0.0</version>
   
   	<name>goldgem</name>
   	<description>Simple framework that implements some basic design patterns for Java projects using MVC</description>
@@ -17,26 +17,23 @@
   	<developers>
         <developer>
           <name>Mariana Azevedo</name>
-          <email>mariana@bsi.ufla.br</email>
-          <timezone>UTC-03:00</timezone>
+          <email>marianaazsantos88@gmail.com</email>
+          <timezone>UTC</timezone>
         </developer>
   	</developers>
   
   	<properties>
-  	
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <mysql.connector.version>8.0.32</mysql.connector.version>
+        <java.version>17</java.version>
         <hibernate.version>5.6.14.Final</hibernate.version>
-        <log4j.version>1.2.17</log4j.version>
-        <joda.time.version>2.12.2</joda.time.version>
+        <log4j2.version>2.17.2</log4j2.version>
+        <h2.version>2.1.214</h2.version>
         <junit.jupiter.version>5.4.0</junit.jupiter.version>
-        <compiler.plugin.version>3.6.1</compiler.plugin.version>
+        <compiler.plugin.version>3.8.1</compiler.plugin.version>
         <surefire.plugin.version>2.22.0</surefire.plugin.version>
         <coveralls.version>4.3.0</coveralls.version>
-        <jacoco.version>0.7.9</jacoco.version>
-		
+        <jacoco.version>0.8.8</jacoco.version>
   	</properties>
   
   	<dependencies>
@@ -46,20 +43,15 @@
 		    <version>${hibernate.version}</version>
 		</dependency>
 		<dependency>
-		    <groupId>mysql</groupId>
-		    <artifactId>mysql-connector-java</artifactId>
-		    <version>${mysql.connector.version}</version>
+		    <groupId>org.apache.logging.log4j</groupId>
+		    <artifactId>log4j-api</artifactId>
+		    <version>${log4j2.version}</version>
 		</dependency>
-		<dependency>
-		    <groupId>log4j</groupId>
-		    <artifactId>log4j</artifactId>
-		    <version>${log4j.version}</version>
-		</dependency>
-		<dependency>
-		    <groupId>joda-time</groupId>
-		    <artifactId>joda-time</artifactId>
-		    <version>${joda.time.version}</version>
-		</dependency>
+	  	<dependency>
+	    	<groupId>org.apache.logging.log4j</groupId>
+	    	<artifactId>log4j-core</artifactId>
+	    	<version>${log4j2.version}</version>
+	  	</dependency>
 		<!-- https://mvnrepository.com/artifact/org.junit.jupiter/junit-jupiter-api -->
 		<dependency>
 		    <groupId>org.junit.jupiter</groupId>
@@ -81,6 +73,12 @@
 		    <version>${junit.jupiter.version}</version>
 		    <scope>test</scope>
 		</dependency>
+		<dependency>
+		    <groupId>com.h2database</groupId>
+		    <artifactId>h2</artifactId>
+		    <version>${h2.version}</version>
+		    <scope>test</scope>
+		</dependency>
    	</dependencies>
 
    	<build>
@@ -91,8 +89,7 @@
             	<artifactId>maven-compiler-plugin</artifactId>
             	<version>${compiler.plugin.version}</version>
            		<configuration>
-                	<source>${java.version}</source>
-                	<target>${java.version}</target>
+                	<release>${java.version}</release>
             	</configuration>
         	</plugin>
         	<plugin>
@@ -111,9 +108,6 @@
 			    <groupId>org.jacoco</groupId>
 			    <artifactId>jacoco-maven-plugin</artifactId>
 			    <version>${jacoco.version}</version>
-			    <configuration>
-			    	<scope>test</scope>
-			    </configuration>
 			    <executions>
 			        <execution>
 		                <id>prepare-agent</id>
