@@ -19,7 +19,7 @@ import javax.persistence.MappedSuperclass;
 public class GenericDTO {
 
 	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private long id;
 	
 	@Column(name = "created_at", nullable = false)
